requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
pytz==2024.1
gunicorn==21.2.0
email-validator==2.1.0.post1
PyJWT==2.10.1
SQLAlchemy==2.0.40
bcrypt==4.1.2
# Dependencias para Google Cloud SQL
google-cloud-sql-connector==1.2.0